moved versioning to modules
