+++ src/main/java/org/betonquest/betonquest/BetonQuest.java
@@ -149,6 +149,7 @@ import org.betonquest.betonquest.item.QuestItemHandler;
 import org.betonquest.betonquest.mechanics.PlayerHider;
 import org.betonquest.betonquest.menu.RPGMenu;
 import org.betonquest.betonquest.modules.logger.LogWatcher;
+import org.betonquest.betonquest.modules.versioning.Updater;
 import org.betonquest.betonquest.notify.ActionBarNotifyIO;
 import org.betonquest.betonquest.notify.AdvancementNotifyIO;
 import org.betonquest.betonquest.notify.BossBarNotifyIO;
@@ -193,7 +194,6 @@ import org.betonquest.betonquest.objectives.TameObjective;
 import org.betonquest.betonquest.objectives.VariableObjective;
 import org.betonquest.betonquest.utils.PlayerConverter;
 import org.betonquest.betonquest.utils.Utils;
-import org.betonquest.betonquest.utils.versioning.Updater;
 import org.betonquest.betonquest.variables.ConditionVariable;
 import org.betonquest.betonquest.variables.GlobalPointVariable;
 import org.betonquest.betonquest.variables.ItemAmountVariable;
@@ -888,7 +888,7 @@ public class BetonQuest extends JavaPlugin {
         new BStatsMetrics(this, metricsSuppliers);
 
         // updater
-        updater = new Updater(this.getDescription().getVersion(), this.getFile());
+        updater = new Updater(config, this.getFile(), this.getDescription().getVersion());
 
         //RPGMenu integration
         rpgMenu = new RPGMenu();
