+++ src/main/java/org/betonquest/betonquest/compatibility/shopkeepers/ShopkeepersIntegrator.java
@@ -5,9 +5,9 @@ import org.betonquest.betonquest.BetonQuest;
 import org.betonquest.betonquest.compatibility.Integrator;
 import org.betonquest.betonquest.exceptions.HookException;
 import org.betonquest.betonquest.exceptions.UnsupportedVersionException;
-import org.betonquest.betonquest.utils.versioning.UpdateStrategy;
-import org.betonquest.betonquest.utils.versioning.Version;
-import org.betonquest.betonquest.utils.versioning.VersionComparator;
+import org.betonquest.betonquest.modules.versioning.UpdateStrategy;
+import org.betonquest.betonquest.modules.versioning.Version;
+import org.betonquest.betonquest.modules.versioning.VersionComparator;
 import org.bukkit.Bukkit;
 import org.bukkit.plugin.Plugin;
