+++ src/main/java/org/betonquest/betonquest/modules/versioning/UpdateStrategy.java
@@ -1,4 +1,4 @@
-package org.betonquest.betonquest.utils.versioning;
+package org.betonquest.betonquest.modules.versioning;
 
 /**
  * Represent different strategies to select which versions are valid to update to.
