+++ src/test/java/org/betonquest/betonquest/modules/versoning/VersionComparatorTest.java
@@ -1,8 +1,8 @@
-package org.betonquest.betonquest.utils.versoning;
+package org.betonquest.betonquest.modules.versoning;
 
-import org.betonquest.betonquest.utils.versioning.UpdateStrategy;
-import org.betonquest.betonquest.utils.versioning.Version;
-import org.betonquest.betonquest.utils.versioning.VersionComparator;
+import org.betonquest.betonquest.modules.versioning.UpdateStrategy;
+import org.betonquest.betonquest.modules.versioning.Version;
+import org.betonquest.betonquest.modules.versioning.VersionComparator;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
